refactor: consolidate data module helper functions
- Create akyo-data-helpers.ts with common extractCategories, extractAuthors, findAkyoById
- Remove duplicate implementations from akyo-data.ts, akyo-data-server.ts, akyo-data-json.ts, akyo-data-kv.ts
- Remove unused/deprecated functions: getAllAttributes, getAllCreators, getDataSourceInfo, getKVStatus, checkKVStatus, clearKVCache, getKVMetadata
- Remove unused exports: getAkyoByIdFromJSON, getAllCategoriesFromJSON, getAllAuthorsFromJSON, getAkyoByIdFromKV, getAllCategoriesFromKV, getAllAuthorsFromKV, getAkyoDataFromKV
- Remove unused AkyoCsvRow interface
- Update knip.json: add dynamic import files to ignore, remove tailwindcss from ignoreDependencies
- Reduces code duplication and improves maintainability
