+++ src/lib/akyo-data-json.ts
@@ -164,62 +164,3 @@ function normalizeAkyoItem(item: unknown): AkyoData {
     avatarUrl: String(raw.avatarUrl || ''),
   };
 }
-
-/**
- * Get single Akyo by ID from JSON data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param id - 4-digit ID (e.g., "0001")
- * @param lang - Language code
- * @returns Single Akyo data or null if not found
- */
-export const getAkyoByIdFromJSON = cache(
-  async (id: string, lang: SupportedLanguage = 'ja'): Promise<AkyoData | null> => {
-    const data = await getAkyoDataFromJSON(lang);
-    return data.find((akyo) => akyo.id === id) || null;
-  }
-);
-
-/**
- * Get all unique categories from JSON data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param lang - Language code
- * @returns Array of unique categories
- */
-export const getAllCategoriesFromJSON = cache(
-  async (lang: SupportedLanguage = 'ja'): Promise<string[]> => {
-    const data = await getAkyoDataFromJSON(lang);
-    const categoriesSet = new Set<string>();
-    
-    data.forEach((akyo) => {
-      const catStr = akyo.category || akyo.attribute || '';
-      const cats = catStr.split(/[、,]/).map((c) => c.trim()).filter(Boolean);
-      cats.forEach((cat) => categoriesSet.add(cat));
-    });
-    
-    return Array.from(categoriesSet).sort();
-  }
-);
-
-/**
- * Get all unique authors from JSON data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param lang - Language code
- * @returns Array of unique authors
- */
-export const getAllAuthorsFromJSON = cache(
-  async (lang: SupportedLanguage = 'ja'): Promise<string[]> => {
-    const data = await getAkyoDataFromJSON(lang);
-    const authorsSet = new Set<string>();
-    
-    data.forEach((akyo) => {
-      const authorStr = akyo.author || akyo.creator || '';
-      const authors = authorStr.split(/[、,]/).map((a) => a.trim()).filter(Boolean);
-      authors.forEach((author) => authorsSet.add(author));
-    });
-    
-    return Array.from(authorsSet).sort();
-  }
-);
