+++ src/lib/akyo-data-kv.ts
@@ -165,80 +165,6 @@ export const getAkyoDataFromKVWithSource = cache(
 );
 
 /**
- * Fetch Akyo data from KV with fallback to JSON
- * Wrapped with React cache() for automatic deduplication within a single request
- * 
- * @deprecated Use getAkyoDataFromKVOnly or getAkyoDataFromKVWithSource for clearer semantics
- * @param lang - Language code (default: 'ja')
- * @returns Array of Akyo data
- */
-export const getAkyoDataFromKV = cache(
-  async (lang: SupportedLanguage = 'ja'): Promise<AkyoData[]> => {
-    const result = await getAkyoDataFromKVWithSource(lang);
-    return result.data;
-  }
-);
-
-/**
- * Get single Akyo by ID from KV data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param id - 4-digit ID (e.g., "0001")
- * @param lang - Language code
- * @returns Single Akyo data or null if not found
- */
-export const getAkyoByIdFromKV = cache(
-  async (id: string, lang: SupportedLanguage = 'ja'): Promise<AkyoData | null> => {
-    const data = await getAkyoDataFromKV(lang);
-    return data.find((akyo) => akyo.id === id) || null;
-  }
-);
-
-/**
- * Get all unique categories from KV data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param lang - Language code
- * @returns Array of unique categories
- */
-export const getAllCategoriesFromKV = cache(
-  async (lang: SupportedLanguage = 'ja'): Promise<string[]> => {
-    const data = await getAkyoDataFromKV(lang);
-    const categoriesSet = new Set<string>();
-    
-    data.forEach((akyo) => {
-      const catStr = akyo.category || akyo.attribute || '';
-      const cats = catStr.split(/[、,]/).map((c) => c.trim()).filter(Boolean);
-      cats.forEach((cat) => categoriesSet.add(cat));
-    });
-    
-    return Array.from(categoriesSet).sort();
-  }
-);
-
-/**
- * Get all unique authors from KV data
- * Wrapped with React cache() for automatic deduplication
- * 
- * @param lang - Language code
- * @returns Array of unique authors
- */
-export const getAllAuthorsFromKV = cache(
-  async (lang: SupportedLanguage = 'ja'): Promise<string[]> => {
-    const data = await getAkyoDataFromKV(lang);
-    const authorsSet = new Set<string>();
-    
-    data.forEach((akyo) => {
-      const authorStr = akyo.author || akyo.creator || '';
-      const authors = authorStr.split(/[、,]/).map((a) => a.trim()).filter(Boolean);
-      authors.forEach((author) => authorsSet.add(author));
-    });
-    
-    return Array.from(authorsSet).sort();
-  }
-);
-
-/**
  * Update KV cache with new data for a single language
  * 
  * WARNING: When updating both languages, use updateKVCacheBoth() instead
@@ -368,83 +294,5 @@ export async function updateKVCacheBoth(
   } catch (error) {
     console.error('[KV] Error in updateKVCacheBoth:', error);
     return result;
-  }
-}
-
-/**
- * Clear all KV cache data
- * Use with caution - removes all cached data
- * 
- * @returns true if successful, false otherwise
- */
-export async function clearKVCache(): Promise<boolean> {
-  const kv = getKVNamespace();
-  
-  if (!kv) {
-    console.warn('[KV] Cannot clear KV: namespace not available');
-    return false;
-  }
-  
-  try {
-    await Promise.all([
-      kv.delete(KV_KEYS.DATA_JA),
-      kv.delete(KV_KEYS.DATA_EN),
-      kv.delete(KV_KEYS.META),
-    ]);
-    console.log('[KV] Cache cleared');
-    return true;
-  } catch (error) {
-    console.error('[KV] Error clearing KV cache:', error);
-    return false;
-  }
-}
-
-/**
- * Get KV cache metadata
- * Useful for debugging and monitoring
- * 
- * @returns Metadata or null if not available
- */
-export async function getKVMetadata(): Promise<KVMetadata | null> {
-  const kv = getKVNamespace();
-  
-  if (!kv) {
-    return null;
-  }
-  
-  try {
-    return await kv.get<KVMetadata>(KV_KEYS.META, { type: 'json' });
-  } catch {
-    return null;
-  }
-}
-
-/**
- * Check if KV cache is available and has data
- * 
- * @returns Object with availability status
- */
-export async function checkKVStatus(): Promise<{
-  available: boolean;
-  hasData: boolean;
-  metadata: KVMetadata | null;
-}> {
-  const kv = getKVNamespace();
-  
-  if (!kv) {
-    return { available: false, hasData: false, metadata: null };
-  }
-  
-  try {
-    const metadata = await kv.get<KVMetadata>(KV_KEYS.META, { type: 'json' });
-    const hasData = metadata !== null && (metadata.countJa > 0 || metadata.countEn > 0);
-    
-    return {
-      available: true,
-      hasData,
-      metadata,
-    };
-  } catch {
-    return { available: true, hasData: false, metadata: null };
   }
 }
